Oh and this

+++ Object3D.h
@@ -13,7 +13,7 @@ public:
 	glm::mat4* Transform();
 
 	virtual void Init() = 0;
-	virtual void Update(float dt) = 0;
+	virtual void Update(double dt) = 0;
 	virtual void Draw() = 0;
 };
